requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
boto3==1.18.22
botocore==1.21.22
cement==2.8.2
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
distlib==0.3.8
distro==1.9.0
Django==3.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-session-timeout==0.1.0
django-storages==1.11.1
django-widget-tweaks==1.5.0
dlib @ file:///C:/django/entrega/proyectoFinal/proyecto/dlib-19.22.1-cp39-cp39-win_amd64.whl#sha256=0980480b21b47e0450752736a4f353915dd3fd565670978822d6b3d78c69de88
docutils==0.16
exceptiongroup==1.2.1
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.15.1
fsspec==2024.6.0
future==0.16.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jmespath==0.10.0
jsonfield==3.1.0
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.2
openai==1.35.3
opencv-python==4.10.0.84
packaging==24.1
pathspec==0.10.1
pillow==10.3.0
pipdeptree==2.23.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.8.6
pyasn1==0.6.0
pydantic==2.7.4
pydantic_core==2.18.4
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==5.3.1
reportlab==4.2.0
requests==2.32.3
rsa==4.7.2
s3transfer==0.5.2
semantic-version==2.8.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.12.0
termcolor==1.1.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.6
virtualenv==20.26.3
wcwidth==0.1.9
